backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.3.1",
    "aiodataloader==0.4.2",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "astunparse==1.6.3",
    "asyncpg==0.30.0",
    "bcrypt>=5.0.0",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "deepface==0.0.95",
    "dependency-injector==4.48.1",
    "ecdsa==0.19.1",
    "eval-type-backport==0.2.2",
    "fastapi==0.116.1",
    "filelock==3.19.1",
    "fire==0.7.1",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "flatbuffers==25.2.10",
    "gast==0.6.0",
    "gdown==5.2.0",
    "google-pasta==0.2.0",
    "graphql-core==3.2.6",
    "greenlet==3.2.4",
    "grpcio==1.74.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "h5py==3.14.0",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "keras==3.11.2",
    "lia-web==0.2.3",
    "libclang==18.1.1",
    "libcst==1.8.2",
    "lz4==4.4.4",
    "markdown==3.8.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.3",
    "mtcnn==1.0.0",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "namex==0.1.0",
    "numpy==2.1.3",
    "opencv-python==4.12.0.88",
    "opt-einsum==3.4.0",
    "optree==0.17.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pathspec==0.12.1",
    "pgvector==0.4.1",
    "pillow==11.3.0",
    "protobuf==5.29.5",
    "pwdlib==0.3.0",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic[email]==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.5",
    "retina-face==0.0.17",
    "rich==14.1.0",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.43",
    "starlette==0.47.2",
    "strawberry-graphql==0.280.0",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.19.1",
    "termcolor==3.1.0",
    "tf-keras==2.19.0",
    "tqdm==4.67.1",
    "typer==0.16.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wireup==2.0.1",
    "wrapt==1.17.3",
    "httpx>=0.28.1",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "FA",
    "PYI",
    "C4",
    "TID",
]
